Changes
Changed the resume so it actually fit within the viewport of the Resume section. Added padding to top and bottom of resume card so it looked neat and more uniform with the rest of the cards in the other sections. Also made background of DOM body black, so when overscrolling or viewing on mobile device there is no more white border.